+++ src/config/axios.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-import { ACCESS_TOKEN, AUTHORIZATION } from "../constants";
+import { ACCESS_TOKEN, AUTHORIZATION } from "../constants/auth";
 
 const instance = axios.create({
   baseURL: process.env.REACT_APP_SERVER_URL,
