+++ src/constants/auth.js
@@ -1,0 +1,2 @@
+export const ACCESS_TOKEN = "accessToken";
+export const AUTHORIZATION = "Authorization";
