+++ src/constants/date.js
@@ -1,0 +1,3 @@
+export const DATE_FORMAT = {
+  YYYY_MM_DD: "yyyy-MM-dd",
+};
