+++ src/constants/diary.js
@@ -1,6 +1,3 @@
-export const ACCESS_TOKEN = "accessToken";
-export const AUTHORIZATION = "Authorization";
-
 export const EFFECT_MODE = "effect";
 export const SCRIPT_MODE = "script";
 
@@ -8,8 +5,4 @@ export const SENTIMENT = {
   ALL: "all",
   POSITIVE: "positive",
   NEGATIVE: "negative",
-};
-
-export const DATE_FORMAT = {
-  YYYY_MM_DD: "yyyy-MM-dd",
 };
